requirements.txt
# Add project pip dependencies here, e.g. flask, cryptography
flask==2.3.0
pycryptodome==3.18.0
flask-login==0.6.2
flask-wtf==1.1.1
email-validator==2.0.0
python-dotenv==1.0.0